unfold: Refactor code resposible for unfolding of one function
